+++ client/App.tsx
@@ -21,9 +21,7 @@ import { useCallback } from "react";
 import { TagState } from "../common/CombatantState";
 import { PendingPrompts } from "./Prompts/PendingPrompts";
 import { CombatFooter } from "./CombatFooter/CombatFooter";
-import { Button } from "./Components/Button";
-import { CombatantCommander } from "./Commands/CombatantCommander";
-import { MultipleCombatantDetails } from "./Combatant/MultipleCombatantDetails";
+import { SelectedCombatants } from "./SelectedCombatants";
 
 /*
  * This file is new as of 05/2020. Most of the logic was extracted from TrackerViewModel.
@@ -239,51 +237,4 @@ function InitiativeListHost(props: { tracker: TrackerViewModel }) {
       />
     </CommandContext.Provider>
   );
-}
-
-function SelectedCombatants(props: { combatantCommander: CombatantCommander }) {
-  const { combatantCommander } = props;
-  const selectedCombatants = useSubscription(
-    combatantCommander.SelectedCombatants
-  );
-
-  if (selectedCombatants.length === 0) {
-    return (
-      <div className="selected-combatant">
-        <h2>No Combatant Selected</h2>
-      </div>
-    );
-  }
-
-  if (selectedCombatants.length === 1) {
-    return (
-      <div className="selected-combatant">
-        <div className="combatant-details__header">
-          <h2>Selected Combatant</h2>
-          <Button
-            fontAwesomeIcon="times"
-            onClick={combatantCommander.Deselect}
-          />
-        </div>
-        <CombatantDetails
-          key={selectedCombatants[0].Combatant.Id}
-          combatantViewModel={selectedCombatants[0]}
-          displayMode="default"
-        />
-      </div>
-    );
-  } else {
-    return (
-      <div className="selected-combatant">
-        <div className="combatant-details__header">
-          <h2>Selected Combatants</h2>
-          <Button
-            fontAwesomeIcon="times"
-            onClick={combatantCommander.Deselect}
-          />
-        </div>
-        <MultipleCombatantDetails combatants={selectedCombatants} />
-      </div>
-    );
-  }
 }
